+++ src/widgets/index.ts
@@ -1,0 +1,38 @@
+import { Application, IPlugin } from '@lumino/application';
+import { Widget } from '@lumino/widgets';
+import { IJupyterWidgetRegistry } from '@jupyter-widgets/base';
+
+import { widgetExports } from '../index-widgets';
+import { MODULE_NAME, MODULE_VERSION } from '../version';
+
+
+
+const EXTENSION_ID = 'jupysql-plugin:plugin';
+
+/**
+ * The widgets plugin.
+ */
+const plugin_widget: IPlugin<Application<Widget>, void> = {
+    id: EXTENSION_ID,
+    requires: [IJupyterWidgetRegistry],
+    activate: activateWidgetExtension,
+    autoStart: true,
+} as unknown as IPlugin<Application<Widget>, void>;
+// the "as unknown as ..." typecast above is solely to support JupyterLab 1
+// and 2 in the same codebase and should be removed when we migrate to Lumino.
+
+/**
+ * Activate the widget extension.
+ */
+function activateWidgetExtension(
+    app: Application<Widget>,
+    registry: IJupyterWidgetRegistry
+): void {
+    registry.registerWidget({
+        name: MODULE_NAME,
+        version: MODULE_VERSION,
+        exports: widgetExports,
+    });
+}
+
+export { plugin_widget }
